Move class `Scope` member implementation into .cpp
