+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/scope.hpp
@@ -15,21 +15,17 @@
 #ifndef OPENSCENARIO_INTERPRETER__SCOPE_HPP_
 #define OPENSCENARIO_INTERPRETER__SCOPE_HPP_
 
-#include <boost/algorithm/string.hpp>
-#include <limits>
 #include <memory>
 #include <openscenario_interpreter/syntax/entity_ref.hpp>
 #include <openscenario_interpreter/syntax/file.hpp>
-#include <scenario_simulator_exception/exception.hpp>
 #include <unordered_map>
 #include <utility>
 #include <vector>
 
 namespace openscenario_interpreter
 {
-struct EnvironmentFrame
+class EnvironmentFrame
 {
-private:
   friend struct Scope;
 
   const std::string scope_name;
@@ -44,166 +40,31 @@ private:
 
   explicit EnvironmentFrame() = default;
 
-  explicit EnvironmentFrame(EnvironmentFrame & parent, const std::string & name)
-  : scope_name(name), parent(&parent)
-  {
-    if (name.empty()) {
-      parent.anonymous_children.push_back(this);
-    } else {
-      parent.named_children.emplace(name, this);
-    }
-  }
+  explicit EnvironmentFrame(EnvironmentFrame &, const std::string &);
 
   explicit EnvironmentFrame(const EnvironmentFrame &) = delete;
 
   explicit EnvironmentFrame(EnvironmentFrame &&) = delete;
 
 public:
-  auto insert(const std::string & name, Element element) -> void
-  {
-    if (name.find(':') != std::string::npos) {
-      THROW_SYNTAX_ERROR("Identifier '", name, "' contains ':'");
-    }
+  auto findElement(const std::string &) const -> Element;
 
-    environments.emplace(name, std::move(element));
-  }
+  auto getQualifiedName() const -> std::string;
 
-  auto findElement(const std::string & name) const -> Element
-  {
-    std::vector<std::string> split;
-    {
-      const char * delim = "::";
-      const std::size_t delim_len = 2;
-
-      std::size_t prev_pos = 0;
-      std::size_t pos = 0;
-      while ((pos = name.find(delim, prev_pos)) != std::string::npos) {
-        split.push_back(name.substr(prev_pos, pos - prev_pos));
-        prev_pos = pos + delim_len;
-      }
-      split.push_back(name.substr(prev_pos, pos));
-    }
-
-    if (split.size() == 1) {
-      return lookupUnqualifiedElement(split.front());
-    } else {
-      auto top_scope = lookupUnqualifiedScope(split.front());
-      if (top_scope) {
-        return lookupQualifiedElement(top_scope, split.begin() + 1, split.end());
-      } else {
-        return Element{};
-      }
-    }
-  }
-
-  auto getQualifiedName() const -> std::string
-  {
-    std::list<const EnvironmentFrame *> ancestors;
-    for (auto * p = this; p != nullptr; p = p->parent) {
-      ancestors.push_back(p);
-    }
-    std::string ret;
-    for (auto it = ancestors.rbegin(); it != ancestors.rend(); ++it) {
-      ret += (it == ancestors.rbegin() ? "" : "::");
-      ret += (*it)->scope_name.empty() ? "{anonymous}" : (*it)->scope_name;
-    }
-    return ret;
-  }
+  auto insert(const std::string &, Element) -> void;
 
 private:
-  auto lookupChildElement(const std::string & name) const -> Element
-  {
-    std::vector<const EnvironmentFrame *> same_level{this};
+  /*  */ auto lookupChildElement(const std::string &) const -> Element;
 
-    while (not same_level.empty()) {
-      std::vector<const EnvironmentFrame *> next_level;
-      std::vector<Element> ret;
+  /*  */ auto lookupChildScope(const std::string &) const -> std::list<const EnvironmentFrame *>;
 
-      for (auto * frame : same_level) {
-        auto range = frame->environments.equal_range(name);
-        for (auto it = range.first; it != range.second; ++it) {
-          ret.push_back(it->second);
-        }
-
-        for (auto * f : frame->anonymous_children) {
-          next_level.push_back(f);
-        }
-      }
-
-      if (ret.size() == 1) {
-        return ret.front();
-      }
-
-      if (ret.size() > 1) {
-        THROW_SYNTAX_ERROR("ambiguous reference to ", std::quoted(name));
-      }
-
-      same_level = std::move(next_level);
-    }
-    return Element{};
-  }
-
-  auto lookupUnqualifiedElement(const std::string & name) const -> Element
-  {
-    for (auto * p = this; p != nullptr; p = p->parent) {
-      auto found = p->lookupChildElement(name);
-      if (found) {
-        return found;
-      }
-    }
-    return Element{};
-  }
-
-  template <typename Iterator>
   static auto lookupQualifiedElement(
-    const EnvironmentFrame * scope, Iterator name_begin, Iterator name_end) -> Element
-  {
-    for (auto iter = name_begin; iter != name_end - 1; ++iter) {
-      auto found = scope->lookupChildScope(*iter);
-      if (found.size() == 1) {
-        scope = found.front();
-      } else if (found.empty()) {
-        return Element{};
-      } else if (found.size() > 1) {
-        THROW_SYNTAX_ERROR("ambiguous reference to ", std::quoted(*iter));
-      }
-    }
+    const EnvironmentFrame *, std::vector<std::string>::iterator,
+    std::vector<std::string>::iterator) -> Element;
 
-    return scope->lookupChildElement(*(name_end - 1));
-  }
+  /*  */ auto lookupUnqualifiedElement(const std::string &) const -> Element;
 
-  auto lookupChildScope(const std::string & name) const -> std::list<const EnvironmentFrame *>
-  {
-    auto range = named_children.equal_range(name);
-    std::list<const EnvironmentFrame *> ret;
-    if (range.first != range.second) {
-      for (auto it = range.first; it != range.second; ++it) {
-        ret.push_back(it->second);
-      }
-    } else {
-      for (auto & child : anonymous_children) {
-        ret.merge(child->lookupChildScope(name));
-      }
-    }
-    return ret;
-  }
-
-  auto lookupUnqualifiedScope(const std::string & name) const -> const EnvironmentFrame *
-  {
-    if (parent == nullptr) {  // this is global scope
-      return name.empty() ? this : nullptr;
-    } else {
-      auto sibling_scope = parent->lookupChildScope(name);
-      if (sibling_scope.size() == 1) {
-        return sibling_scope.front();
-      } else if (sibling_scope.size() > 1) {
-        THROW_SYNTAX_ERROR("ambiguous reference to ", name);
-      } else if (sibling_scope.empty() && parent) {
-        return parent->lookupUnqualifiedScope(name);
-      }
-      return nullptr;
-    }
-  }
+  /*  */ auto lookupUnqualifiedScope(const std::string &) const -> const EnvironmentFrame *;
 };
 
 class Scope
@@ -216,7 +77,7 @@ class Scope
 
     std::unordered_map<std::string, Element> entities;  // ScenarioObject or EntitySelection
 
-    explicit GlobalEnvironment(const File::Path & pathname) : pathname(pathname) {}
+    explicit GlobalEnvironment(const File::Path &);
 
     auto entityRef(const EntityRef &) const -> Element;  // TODO: RETURN ScenarioObject TYPE!
 
@@ -232,43 +93,29 @@ public:
 
   explicit Scope() = delete;
 
-  explicit Scope(const File::Path & pathname)
-  : frame(new EnvironmentFrame()), global_environment(std::make_shared<GlobalEnvironment>(pathname))
-  {
-  }
+  explicit Scope(const File::Path &);
 
 private:
-  explicit Scope(
-    const Scope & parent, const std::string & name, const std::shared_ptr<EnvironmentFrame> & frame)
-  : frame(frame), global_environment(parent.global_environment), name(name), actors(parent.actors)
-  {
-  }
+  explicit Scope(const Scope &, const std::string &, const std::shared_ptr<EnvironmentFrame> &);
 
 public:
   Scope(const Scope &) = default;  // NOTE: shallow copy
 
   Scope(Scope &&) noexcept = default;
 
-  auto global() const -> const auto & { return *global_environment; }
+  auto findElement(const std::string & name_) const -> Element;
 
-  auto global() -> auto & { return *global_environment; }
+  auto global() const -> const GlobalEnvironment &;
 
-  auto localScope() const noexcept -> const auto & { return *this; }
+  auto global() -> GlobalEnvironment &;
 
-  auto localScope() noexcept -> auto & { return *this; }
+  auto localScope() const noexcept -> const Scope &;
 
-  auto makeChildScope(const std::string & name) const
-  {
-    return Scope(
-      *this, name, std::shared_ptr<EnvironmentFrame>(new EnvironmentFrame(*frame, name)));
-  }
+  auto localScope() noexcept -> Scope &;
 
-  auto insert(const std::string & name_, const Element & element)
-  {
-    return frame->insert(name_, element);
-  }
+  auto makeChildScope(const std::string &) const -> Scope;
 
-  auto findElement(const std::string & name_) const { return frame->findElement(name_); }
+  auto insert(const std::string & name_, const Element & element) -> void;
 };
 }  // namespace openscenario_interpreter
